google_genai_hello_world/pyproject.toml
[project]
name = "google-genai-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
